Experimental file backed index.

+++ config/config.go
@@ -9,7 +9,7 @@ type AppConfig struct {
 // FileStoreConfig contains the configuration for storing data in files on disk.
 type FileStoreConfig struct {
 	Path           string
-	SystemsPerFile int64
+	SystemsPerFile int
 }
 
 // AccessControlConfig defines the requirements for an incoming request to be handled.
